requirements.txt
aiohttp>=3.9.0,<4.0.0
pywebview>=5.0,<6.0
